BACKEND: Move resolv.conf watching to the backends
Moved from the monitor, which no longer reads the file neither
notifies the backends they must read it.

The file is now monitored by the backends and read immediately
after start watching it.

Resolves: https://github.com/SSSD/sssd/issues/6383
